+++ tests/test_tools.py
@@ -334,10 +334,16 @@ def test_add_obs_date(
     ])
 
 
+@pytest.mark.parametrize(
+    "write",
+    [(True), (False)],
+    ids=("write", "no-write")
+)
 def test_gen_mocs_image(
         dummy_fits_open: fits.HDUList,
         dummy_load_fields_file: pd.DataFrame,
         dummy_moc: MOC,
+        write: bool,
         mocker) -> None:
     """
     Tests the generation of a MOC and STMOC for a single fits file
@@ -347,6 +353,7 @@ def test_gen_mocs_image(
             FITS file.
         dummy_load_fields_file: The dummy fields file.
         dummy_moc: The dummy MOC object representing an open MOC.
+        write: Whether to test the write to file or not.
         mocker: The pytest mock mocker object.
 
     Returns:
@@ -402,13 +409,14 @@ def test_gen_mocs_image(
     moc_file = fits_file.replace('.fits', '.moc.fits')
     stmoc_file = fits_file.replace('.fits', '.stmoc.fits')
 
-    moc, stmoc = vtt.gen_mocs_image(fits_file)
+    moc, stmoc = vtt.gen_mocs_image(fits_file, write=write)
 
-    mocker_moc_write.assert_called_once_with(Path(moc_file),
-                                             overwrite=True)
+    if write:
+        mocker_moc_write.assert_called_once_with(Path(moc_file),
+                                                 overwrite=True)
 
-    mocker_stmoc_write.assert_called_once_with(Path(stmoc_file),
-                                               overwrite=True)
+        mocker_stmoc_write.assert_called_once_with(Path(stmoc_file),
+                                                   overwrite=True)
 
     assert stmoc.max_time.jd == end.jd
     assert stmoc.min_time.jd == start.jd
@@ -459,6 +467,66 @@ def test_gen_mocs_epoch(dummy_moc: MOC,
     mocker_stmoc_write.assert_has_calls(stmoc_calls)
 
 
+def test_gen_mocs_epoch_outdir_failure(mocker) -> None:
+    """
+    Tests the generation of all MOCs and STMOCs for a single epoch.
+    Also tests the update of the full STMOC.
+
+    Args:
+        mocker: The pytest mock mocker object.
+
+    Returns:
+        None
+    """
+
+    outdir = '/path/to/outdir'
+
+    mocker_exists = mocker.patch('vasttools.tools.Path.exists',
+                                 return_value=False
+                                 )
+
+    with pytest.raises(Exception) as excinfo:
+        vtt.gen_mocs_epoch('1',
+                           '',
+                           '',
+                           outdir=outdir
+                           )
+
+    exc_str = "{} does not exist".format(outdir)
+
+    assert str(excinfo.value) == exc_str
+
+
+def test_gen_mocs_epoch_stmoc_failure(mocker) -> None:
+    """
+    Tests the generation of all MOCs and STMOCs for a single epoch.
+    Also tests the update of the full STMOC.
+
+    Args:
+        mocker: The pytest mock mocker object.
+
+    Returns:
+        None
+    """
+
+    base_stmoc = '/path/to/stmoc.fits'
+
+    mocker_exists = mocker.patch('vasttools.tools.Path.exists',
+                                 side_effect=[True, False]
+                                 )
+
+    with pytest.raises(Exception) as excinfo:
+        vtt.gen_mocs_epoch('1',
+                           '',
+                           '',
+                           epoch_path='.',
+                           base_stmoc=base_stmoc
+                           )
+
+    exc_str = "{} does not exist".format(base_stmoc)
+    assert str(excinfo.value) == exc_str
+
+
 def test__set_epoch_path_failure(mocker) -> None:
     """
     Tests the set_epoch_path function when `VAST_DATA_DIR` has not been set.
@@ -475,7 +543,7 @@ def test__set_epoch_path_failure(mocker) -> None:
 
     with pytest.raises(Exception) as excinfo:
         vtt._set_epoch_path('1')
-    print(str(excinfo.value))
+
     assert str(excinfo.value).startswith(
         "The path to the requested epoch could not be determined!"
     )